Add social media fields
